Support MultipleException in StepTemplate::Exception
- Integrate StepTemplate::Exception and StepTemplate::Source
- Convert slim to erb
