+++ lib/turnip_formatter/step_template/exception.html.erb
@@ -1,0 +1,19 @@
+<% exceptions.each do |e| %>
+  <div class="step_exception">
+    <dl>
+      <dt>Failure:</dt>
+      <dd>
+        <pre><%= ERB::Util.h(e[:message]) %></pre>
+      </dd>
+      <dt>Backtrace:</dt>
+      <dd>
+        <ol>
+          <% e[:backtrace].each do |line| %>
+            <li><%= ERB::Util.h(line) %></li>
+          <% end %>
+        </ol>
+      </dd>
+    </dl>
+  </div>
+  <pre class="source"><code class="ruby"><%= e[:code] %></code></pre>
+<% end %>
