+++ lib/turnip_formatter/step_template/exception.rb
@@ -1,5 +1,6 @@
 require 'turnip_formatter/step_template/base'
-require 'slim'
+require 'erb'
+require 'ostruct'
 
 module TurnipFormatter
   module StepTemplate
@@ -21,20 +22,46 @@ module TurnipFormatter
               margin-top: 1em;
               margin-left: 1em;
           }
+
+          pre.source {
+              font-size: 12px;
+              font-family: monospace;
+              background-color: #073642;
+              color: #dddddd;
+          }
+
+          pre.source code.ruby {
+              padding: 0.1em 0 0.2em 0;
+          }
+
+          pre.source code.ruby .linenum {
+              width: 75px;
+              color: #fffbd3;
+              padding-right: 1em;
+          }
+
+          pre.source code.ruby .offending {
+              background-color: gray;
+          }
         EOS
       end
 
       #
-      # @param  [RSpec::Core::Example]  example
+      # @param  [TurnipFormatter::Resource::Step::Failure]  step
       #
-      def build_failed(example)
-        exception = example.exception
+      def build_failed(step)
+        datas = step.exceptions.map do |e|
+          backtrace = formatted_backtrace(step.example, e)
+          code = extractor.snippet([backtrace.first])
 
-        if example.exception.is_a?(RSpec::Core::MultipleExceptionError)
-          exception = example.exception.all_exceptions.first
+          {
+            code: code,
+            message: e.to_s,
+            backtrace: backtrace,
+          }
         end
 
-        build(exception.to_s, formatted_backtrace(example, exception))
+        render(exceptions: datas)
       end
 
       #
@@ -46,12 +73,28 @@ module TurnipFormatter
 
       private
 
-      def build(message, backtrace)
-        template_step_exception.render(Object.new, { message: message, backtrace: backtrace })
+      def render(args)
+        view.result(OpenStruct.new(args).instance_eval { binding })
       end
 
-      def template_step_exception
-        @template_step_exception ||= Slim::Template.new(File.dirname(__FILE__) + "/exception.slim")
+      def view
+        @view ||= ::ERB.new(File.read(File.dirname(__FILE__) + '/exception.html.erb'))
+      end
+
+      def extractor
+        @extractor ||= begin
+                         # RSpec 3.4
+                         require 'rspec/core/formatters/html_snippet_extractor'
+                         ::RSpec::Core::Formatters::HtmlSnippetExtractor.new
+                       rescue LoadError
+                         # RSpec 3.3 or earlier
+                         require 'rspec/core/formatters/snippet_extractor'
+                         ::RSpec::Core::Formatters::SnippetExtractor.new
+                       end
+      end
+
+      def formatted_backtrace(example, exception = nil)
+        RSpec::Core::Formatters::TurnipFormatter.formatted_backtrace(example, exception)
       end
     end
   end
