--- lib/turnip_formatter/step_template/exception.slim
@@ -1,10 +1,0 @@
-div.step_exception
-  dl
-    dt Failure:
-    dd
-      pre= message
-    dt Backtrace:
-    dd
-      ol
-       - backtrace.each do |line|
-         li= line
